refactor(install-dynamic-plugins): extract protocol constants and predicates
Move DOCKER_PROTO and OCI_PROTO out of types.ts into a dedicated
protocols.ts module. Add predicate functions (isOciUrl, isDockerUrl,
isHttpUrl, isLocalPath) to replace scattered startsWith checks, using
names that are accurate regardless of whether the caller is checking
a plugin package spec or an image reference.
